+++ inc/Admin/Options.php
@@ -12,10 +12,23 @@ namespace ImageConverterWebP\Admin;
 
 class Options {
 	/**
+	 * The Form.
+	 *
+	 * This array defines every single aspect of the
+	 * Form displayed on the Admin options page.
+	 *
+	 * @since 1.2.0
+	 *
+	 * @var mixed[]
+	 */
+	public static array $form;
+
+	/**
 	 * Define custom static method for calling
 	 * dynamic methods for e.g. Options::get_page_title().
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Invoke init() call here.
 	 *
 	 * @param string  $method Method name.
 	 * @param mixed[] $args   Method args.
@@ -23,32 +36,36 @@ class Options {
 	 * @return string|mixed[]
 	 */
 	public static function __callStatic( $method, $args ) {
+		static::init();
+
 		$keys = substr( $method, strpos( $method, '_' ) + 1 );
 		$keys = explode( '_', $keys );
 
 		$value = '';
 
 		foreach ( $keys as $key ) {
-			$value = empty( $value ) ? ( self::FORM[ $key ] ?? '' ) : ( $value[ $key ] ?? '' );
+			$value = empty( $value ) ? ( static::$form[ $key ] ?? '' ) : ( $value[ $key ] ?? '' );
 		}
 
 		return $value;
 	}
 
 	/**
-	 * The Form.
-	 *
-	 * This array defines every single aspect of the
-	 * Form displayed on the Admin options page.
+	 * Set up Form.
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Set up form using static methods.
+	 *
+	 * @return void
 	 */
-	public const FORM = [
-		'page'   => self::FORM_PAGE,
-		'notice' => self::FORM_NOTICE,
-		'fields' => self::FORM_FIELDS,
-		'submit' => self::FORM_SUBMIT,
-	];
+	public static function init(): void {
+		static::$form = [
+			'page'   => static::get_form_page(),
+			'notice' => static::get_form_notice(),
+			'fields' => static::get_form_fields(),
+			'submit' => static::get_form_submit(),
+		];
+	}
 
 	/**
 	 * Form Page.
@@ -57,13 +74,24 @@ class Options {
 	 * summary, slug and option name.
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Make strings translatable.
+	 *
+	 * @return mixed[]
 	 */
-	public const FORM_PAGE = [
-		'title'   => 'Image Converter for WebP',
-		'summary' => 'Convert your WordPress JPG/PNG images to WebP formats during runtime.',
-		'slug'    => 'image-converter-webp',
-		'option'  => 'icfw',
-	];
+	public static function get_form_page(): array {
+		return [
+			'title'   => esc_html__(
+				'Image Converter for WebP',
+				'image-converter-webp'
+			),
+			'summary' => esc_html__(
+				'Convert your WordPress JPG/PNG images to WebP formats during runtime.',
+				'image-converter-webp'
+			),
+			'slug'    => 'image-converter-webp',
+			'option'  => 'icfw',
+		];
+	}
 
 	/**
 	 * Form Submit.
@@ -72,18 +100,23 @@ class Options {
 	 * button name & label and nonce params.
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Make strings translatable.
+	 *
+	 * @return mixed[]
 	 */
-	public const FORM_SUBMIT = [
-		'heading' => 'Actions',
-		'button'  => [
-			'name'  => 'icfw_save_settings',
-			'label' => 'Save Changes',
-		],
-		'nonce'   => [
-			'name'   => 'icfw_settings_nonce',
-			'action' => 'icfw_settings_action',
-		],
-	];
+	public static function get_form_submit(): array {
+		return [
+			'heading' => esc_html__( 'Actions', 'image-converter-webp' ),
+			'button'  => [
+				'name'  => 'icfw_save_settings',
+				'label' => esc_html__( 'Save Changes', 'image-converter-webp' ),
+			],
+			'nonce'   => [
+				'name'   => 'icfw_settings_nonce',
+				'action' => 'icfw_settings_action',
+			],
+		];
+	}
 
 	/**
 	 * Form Fields.
@@ -92,57 +125,62 @@ class Options {
 	 * each group block and controls.
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Make strings translatable.
+	 *
+	 * @return mixed[]
 	 */
-	public const FORM_FIELDS = [
-		'icfw_conv_options' => [
-			'heading'  => 'Conversion Options',
-			'controls' => [
-				'quality'   => [
-					'control'     => 'text',
-					'placeholder' => '50',
-					'label'       => 'Conversion Quality',
-					'summary'     => 'e.g. 75',
-				],
-				'converter' => [
-					'control' => 'select',
-					'label'   => 'WebP Engine',
-					'summary' => 'e.g. Imagick',
-					'options' => [
-						'gd'      => 'GD',
-						'cwebp'   => 'CWebP',
-						'ffmpeg'  => 'FFMPeg',
-						'imagick' => 'Imagick',
-						'gmagick' => 'Gmagick',
+	public static function get_form_fields() {
+		return [
+			'icfw_conv_options' => [
+				'heading'  => esc_html__( 'Conversion Options', 'image-converter-webp' ),
+				'controls' => [
+					'quality'   => [
+						'control'     => esc_attr( 'text' ),
+						'placeholder' => esc_attr__( '50', 'image-converter-webp' ),
+						'label'       => esc_html__( 'Conversion Quality', 'image-converter-webp' ),
+						'summary'     => esc_html__( 'e.g. 75', 'image-converter-webp' ),
+					],
+					'converter' => [
+						'control' => esc_attr( 'select' ),
+						'label'   => esc_attr__( 'WebP Engine', 'image-converter-webp' ),
+						'summary' => esc_html__( 'e.g. Imagick', 'image-converter-webp' ),
+						'options' => [
+							'gd'      => 'GD',
+							'cwebp'   => 'CWebP',
+							'ffmpeg'  => 'FFMPeg',
+							'imagick' => 'Imagick',
+							'gmagick' => 'Gmagick',
+						],
 					],
 				],
 			],
-		],
-		'icfw_img_options'  => [
-			'heading'  => 'Image Options',
-			'controls' => [
-				'upload'    => [
-					'control' => 'checkbox',
-					'label'   => 'Convert Images on Upload',
-					'summary' => 'This is useful for new images.',
-				],
-				'page_load' => [
-					'control' => 'checkbox',
-					'label'   => 'Convert Images on Page Load',
-					'summary' => 'This is useful for existing images.',
-				],
-			],
-		],
-		'icfw_log_options'  => [
-			'heading'  => 'Log Options',
-			'controls' => [
-				'logs' => [
-					'control' => 'checkbox',
-					'label'   => 'Log errors for Failed Conversions',
-					'summary' => 'Enable this option to log errors.',
+			'icfw_img_options'  => [
+				'heading'  => esc_html__( 'Image Options', 'image-converter-webp' ),
+				'controls' => [
+					'upload'    => [
+						'control' => esc_attr( 'checkbox' ),
+						'label'   => esc_html__( 'Convert Images on Upload', 'image-converter-webp' ),
+						'summary' => esc_html__( 'This is useful for new images.', 'image-converter-webp' ),
+					],
+					'page_load' => [
+						'control' => esc_attr( 'checkbox' ),
+						'label'   => esc_html__( 'Convert Images on Page Load', 'image-converter-webp' ),
+						'summary' => esc_html__( 'This is useful for existing images.', 'image-converter-webp' ),
+					],
 				],
 			],
-		],
-	];
+			'icfw_log_options'  => [
+				'heading'  => 'Log Options',
+				'controls' => [
+					'logs' => [
+						'control' => esc_attr( 'checkbox' ),
+						'label'   => esc_html__( 'Log errors for Failed Conversions', 'image-converter-webp' ),
+						'summary' => esc_html__( 'Enable this option to log errors.', 'image-converter-webp' ),
+					],
+				],
+			],
+		];
+	}
 
 	/**
 	 * Form Notice.
@@ -151,8 +189,13 @@ class Options {
 	 * text displayed on save.
 	 *
 	 * @since 1.1.2
+	 * @since 1.2.0 Make strings translatable.
+	 *
+	 * @return mixed[]
 	 */
-	public const FORM_NOTICE = [
-		'label' => 'Settings Saved.',
-	];
+	public static function get_form_notice(): array {
+		return [
+			'label' => esc_html__( 'Settings Saved.', 'image-converter-webp' ),
+		];
+	}
 }