+++ inc/Services/Logger.php
@@ -22,7 +22,7 @@ class Logger extends Service implements Kernel {
 	 * @return void
 	 */
 	public function register(): void {
-		add_action( 'icfw_convert', [ $this, 'add_webp_meta_to_attachment' ], 10, 2 );
+		add_action( 'icfw_convert', [ $this, 'add_logs_for_webp_conversions' ], 10, 2 );
 	}
 
 	/**
@@ -33,17 +33,14 @@ class Logger extends Service implements Kernel {
 	 *
 	 * @since 1.0.2
 	 * @since 1.1.0 Moved to Logger class.
+	 * @since 1.2.0 Moved meta data logic to MetaData class.
 	 *
 	 * @param string|\WP_Error $webp          WebP's relative path.
 	 * @param int              $attachment_id Image ID.
 	 *
 	 * @return void
 	 */
-	public function add_webp_meta_to_attachment( $webp, $attachment_id ): void {
-		if ( ! is_wp_error( $webp ) && ! get_post_meta( $attachment_id, 'icfw_img', true ) ) {
-			update_post_meta( $attachment_id, 'icfw_img', $webp );
-		}
-
+	public function add_logs_for_webp_conversions( $webp, $attachment_id ): void {
 		if ( ! icfw_get_settings( 'logs' ) ) {
 			return;
 		}