+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.2.0
+* Resolve issue with undefined array keys in Main service.
+* Serve WebP images in WP media library.
+* Refactor Form & Option classes.
+* Make strings translatable across plugin.
+* Fix failing tests, add new tests.
+* Update README notes.
+* Tested up to WP 6.7.1.
+
 ## 1.1.2
 * Refactor Admin page, make extensible with new classes.
 * Add new custom filter `icfw_form_fields`.